+++ projects/ng-snotify/src/lib/decorators/set-toast-type.decorator.ts
@@ -1,14 +1,14 @@
-import { SnotifyType } from '../types/snotify.type';
+import { SnotifyTypeType } from '../types/snotify-type.type';
 import { Snotify } from '../interfaces/snotify.interface';
 
 /**
  * Defines toast style depending on method name
  * @param target any
- * @param propertyKey SnotifyType
+ * @param propertyKey SnotifyTypeType
  * @param descriptor PropertyDescriptor
  * @returns value: ((...args: any[]) => any)
  */
-export function SetToastType(target: any, propertyKey: SnotifyType, descriptor: PropertyDescriptor) {
+export function SetToastType(target: any, propertyKey: SnotifyTypeType, descriptor: PropertyDescriptor) {
   return {
     value(...args: any[]) {
       (args[0] as Snotify).config = {