+++ projects/ng-snotify/src/lib/types/snotify-position.type.ts
@@ -1,0 +1,13 @@
+/**
+ * Toast position types
+ */
+export type SnotifyPositionType =
+  | 'leftTop'
+  | 'leftCenter'
+  | 'leftBottom'
+  | 'rightTop'
+  | 'rightCenter'
+  | 'rightBottom'
+  | 'centerTop'
+  | 'centerCenter'
+  | 'centerBottom';