+++ projects/ng-snotify/src/lib/types/snotify-type.type.ts
@@ -1,0 +1,1 @@
+export type SnotifyTypeType = 'simple' | 'success' | 'error' | 'warning' | 'info' | 'async' | 'confirm' | 'prompt';